refactor(web): playground file stractur
