+++ apps/web/src/hooks/index.ts
@@ -34,3 +34,4 @@ export * from './useVariablesManager';
 export * from './useVercelIntegration';
 export * from './useVercelParams';
 export * from './useApiKey';
+export * from './useContainer';
