+++ apps/web/src/pages/playground/onboarding/PlaygroundTourGuide.tsx
