+++ apps/web/src/pages/playground/onboarding/PlaygroundWorkflowComponent.tsx
