+++ apps/web/src/pages/playground/onboarding/TitleBarWrapper.tsx
